adds custom assert functions, fixes many "function not returning" warnings, adds an enum-switch specific warning
MSVC treats enums as their underlying types (as c++ standard does apparently). This causes a distrust on switch statements that cover all the enum values, because a static_cast might skip all return statements.

For such cases adding a default branch is bad because that would silence the newly added warning if switch does not cover all the values of the enum. Current solution is to put an assert_enum_out_of_range() after the switch.

The new warning also catches cases where a switch on GLenum does not cover all the values (lol). For these cases, a default branch silences the warning. default: assert_case_not_handled() should catch if that happens.